hack: detect incomplete embedding in e2e.test
The e2e.test binary is meant to be self-contained and must embed all
files that it depends on. For example, it gets run outside of the
Kubernetes source tree during conformance testing. The mechanism
for loading files from the current directory (required for
user-configurable files e.g. when running the storage tests against
an out-of-tree CSI driver) can hide missing embedded files.

To detect missing embedding, we change the directory here when running
in a CI environment.