DRA kubelet: give DRA drivers a 5 minute grace period for updates
When doing a rolling update of a DaemonSet, first the old pod gets stopped and
then the new one is started. This causes the kubelet to remove all
ResourceSlices directly after removal and forces the new pod to recreate all of
them.

Now the kubelet waits 5 minutes before it deletes ResourceSlices. If a new
driver registers during that period, nothing is done at all. The new driver
finds the existing ResourceSlices and only needs to update them if something
changed.

The downside is that if the driver gets removed permanently, this creates a 5
minute time window where pods might still get scheduled to the node although
the driver is not going to run there anymore and thus the pods will be stuck.